chore: fmt + clippy + test sweep across all review-driven commits
Final pass after the nine Unit-aligned commits:

* cargo fmt --all (one wide reformat of every file the review
  touched).
* clippy fixups: removed unnecessary u32-as-u32 casts in lockfile
  symlink check; needless_borrow on `segment(1)` in supervisor log
  rotation; identity-op leftover in tui logs `set_tail`; shadow_unrelated
  in `cli/stop` test, `cli/logs` follow path, and `cli/resolve`'s
  parse_running_row.
* Built-in unused-import cleanups now that `lib.rs` no longer carries
  the crate-wide `#![allow(dead_code)]`: BTreeMap / Presets in
  `ipc/methods`, ModelSource / Quant relocated to the `tests` module
  in `discovery/catalog`, `serde_json::json` in `cli/stop`,
  `serde_json::Value` cfg-gated in `gpu/metal` (with the `parse` and
  helper functions also gated to `cfg(any(target_os = "macos", test))`
  so the Linux build doesn't carry dead Apple-specific code).
* `_test_sleep` IPC method retained behind `test-fixtures`.
* `flatten_state` lifted out of `status_response` so `stop_model` and
  `stop_all` responses share the same `{state: {state: "..."}}` shape.
  Error variants carry the cause as a sibling `cause` field instead of
  serde's tagged-enum content.
* Bulk ctor-site fixups in `tests/cli_integration_test.rs` to plumb
  the new `json` / `quiet` / `grace_secs` fields through every
  StartArgs / StopArgs / LogsArgs / FavoritesAction / PresetsAction
  literal.
* `save_does_not_follow_symlink_at_tmp_path` (renamed) now asserts
  the actual invariant: the planted symlink at the tmp path is
  unlinked and replaced, and the victim file is never truncated.
  Previous version expected an `Err` return that the implementation
  doesn't (and shouldn't) produce.
* `list_json_wraps_rows_in_models_object_with_documented_keys` +
  `list_json_empty_catalog_returns_empty_models_array` updated to
  match the new wrapped shape from Unit 8.
* sysinfo API call sites migrated from the deprecated
  `refresh_process` to `refresh_processes_specifics`.

Final state: `cargo fmt --all -- --check` clean, `cargo clippy
--all-targets --features test-fixtures -- -D warnings` clean, `cargo
test --features test-fixtures --workspace` → 459 passed, 0 failed.

+++ src/daemon/lockfile.rs
@@ -137,8 +137,8 @@ pub fn acquire(state_dir: &Path) -> Result<AcquireOutcome, LockfileError> {
   #[cfg(unix)]
   {
     let meta = file.metadata()?;
-    let mode = meta.mode() & libc::S_IFMT as u32;
-    if mode != libc::S_IFREG as u32 {
+    let mode = meta.mode() & libc::S_IFMT;
+    if mode != libc::S_IFREG {
       return Err(LockfileError::CorruptLockfile(
         path.clone(),
         format!("not a regular file (mode {mode:o})"),
